algebra properties simple refactor
