+++ pkg/base/algebra/properties/structure.go
@@ -1,18 +1,22 @@
 package properties
 
 import (
-	"github.com/bronlabs/bron-crypto/pkg/base/algebra"
 	"pgregory.net/rapid"
 )
 
-type Carrier[S, E any] struct {
+type Structure interface {
+	Name() string
+}
+type Element any
+
+type Carrier[S Structure, E Element] struct {
 	Value S
 	Dist  *rapid.Generator[E]
 }
 
 type Carrier2[
-	S1 algebra.Structure[E1], S2 algebra.Structure[E2],
-	E1 algebra.Element[E1], E2 algebra.Element[E2],
+	S1, S2 Structure,
+	E1, E2 Element,
 ] struct {
 	First  *Carrier[S1, E1]
 	Second *Carrier[S2, E2]
